[JAMES-4145] removes invalid dependency to postgres-common-guice
postgres-common-guice should not depend on james-server-guice-distributed ( since that pulls cassandra, opensearch, rabbitmq,... ) when apps can legitimately use the postgres backend without all this.

removing that dependency led to :
- discovering a couple suspicious modules in postgres-common-guice which depend on rabbit and cassandra respectively
- dicovering missing a missing direct dependency in mailbox postgres

The modules were moved to the app itself for now but PostgresDLPConfigurationStoreModule is really suspicious.
