+++ server/apps/migration/core-data-jpa-to-pg/pom.xml
@@ -80,6 +80,10 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
+            <artifactId>james-server-guice-distributed</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${james.groupId}</groupId>
             <artifactId>james-server-guice-sieve-postgres</artifactId>
         </dependency>
         <dependency>
