--- server/container/guice/postgres-common/src/main/java/org/apache/james/modules/data/PostgresDLPConfigurationStoreModule.java
@@ -1,44 +1,0 @@
-/****************************************************************
- * Licensed to the Apache Software Foundation (ASF) under one   *
- * or more contributor license agreements.  See the NOTICE file *
- * distributed with this work for additional information        *
- * regarding copyright ownership.  The ASF licenses this file   *
- * to you under the Apache License, Version 2.0 (the            *
- * "License"); you may not use this file except in compliance   *
- * with the License.  You may obtain a copy of the License at   *
- *                                                              *
- *   http://www.apache.org/licenses/LICENSE-2.0                 *
- *                                                              *
- * Unless required by applicable law or agreed to in writing,   *
- * software distributed under the License is distributed on an  *
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
- * KIND, either express or implied.  See the License for the    *
- * specific language governing permissions and limitations      *
- * under the License.                                           *
- ****************************************************************/
-
-package org.apache.james.modules.data;
-
-import org.apache.james.dlp.api.DLPConfigurationStore;
-import org.apache.james.dlp.eventsourcing.EventSourcingDLPConfigurationStore;
-import org.apache.james.dlp.eventsourcing.cassandra.DLPConfigurationModules;
-import org.apache.james.eventsourcing.Event;
-import org.apache.james.eventsourcing.eventstore.dto.EventDTO;
-import org.apache.james.eventsourcing.eventstore.dto.EventDTOModule;
-
-import com.google.inject.AbstractModule;
-import com.google.inject.Scopes;
-import com.google.inject.TypeLiteral;
-import com.google.inject.multibindings.Multibinder;
-
-public class PostgresDLPConfigurationStoreModule extends AbstractModule {
-
-    @Override
-    protected void configure() {
-        bind(EventSourcingDLPConfigurationStore.class).in(Scopes.SINGLETON);
-        bind(DLPConfigurationStore.class).to(EventSourcingDLPConfigurationStore.class);
-        Multibinder<EventDTOModule<? extends Event, ? extends EventDTO>> eventDTOModuleBinder = Multibinder.newSetBinder(binder(), new TypeLiteral<>() {});
-        eventDTOModuleBinder.addBinding().toInstance(DLPConfigurationModules.DLP_CONFIGURATION_STORE);
-        eventDTOModuleBinder.addBinding().toInstance(DLPConfigurationModules.DLP_CONFIGURATION_CLEAR);
-    }
-}
